Fix old time todo (Function tags)
Function tags are different from callback tags because in the function tags you receive all the content between the start tag and the end tag to your function (You can decide what content you want to return). A callback tag is a simple tag that call a function and return to your function the parameters inside it.

Functions tag included:
-{@userLoggedIn=start@} - {@userLoggedIn=end@}
-{@userNotLoggedIn=start@} - {@userNotLoggedIn=end@}
---These functions are useful in an HTML theme where you can hide some content or show to logged in/out users

Callback tags included:
{@hidePageIfLogged@} - Hide the page if logged
{@hidePageIfNotLogged@} - Hide the page if not logged